Monthly Medicare Australia prescription data

PBS is a monthly tsibble with two values:

Scripts:	Total number of scripts

Cost:	Cost of the scripts in $AUD

Format

Time series of class tsibble

Source

Medicare Australia

Details

The data is disaggregated using four keys:

Concession:	Concessional scripts are given to pensioners, unemployed, dependents, and other card holders

Type:	Co-payments are made until an individual's script expenditure hits a threshold ($290.00 for concession, $1141.80 otherwise). Safety net subsidies are provided to individuals exceeding this amount.

ATC1:	Anatomical Therapeutic Chemical index (level 1)

ATC2:	Anatomical Therapeutic Chemical index (level 2)

Examples

library(tsibble)

#> 
#> Attaching package: ‘tsibble’

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, union

PBS

#> # A tsibble: 67,596 x 9 [1M]
#> # Key:       Concession, Type, ATC1, ATC2 [336]
#>       Month Concession  Type   ATC1  ATC1_desc   ATC2  ATC2_desc   Scripts  Cost
#>       <mth> <chr>       <chr>  <chr> <chr>       <chr> <chr>         <dbl> <dbl>
#>  1 1991 Jul Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   18228 67877
#>  2 1991 Aug Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   15327 57011
#>  3 1991 Sep Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   14775 55020
#>  4 1991 Oct Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   15380 57222
#>  5 1991 Nov Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   14371 52120
#>  6 1991 Dec Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   15028 54299
#>  7 1992 Jan Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   11040 39753
#>  8 1992 Feb Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   15165 54405
#>  9 1992 Mar Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   16898 61108
#> 10 1992 Apr Concession… Co-pa… A     Alimentary… A01   STOMATOLOG…   18141 65356
#> # … with 67,586 more rows