Passenger numbers on Ansett airline flights

The data features a major pilots' industrial dispute which results in some weeks having zero passengers. There were also at least two changes in the definitions of passenger classes.

Format

Time series of class tsibble

Source

Ansett Airlines (which no longer exists).

Details

ansett is a weekly tsibble with one value:

Passengers:	Total air passengers travelling with Ansett

Each series is uniquely identified using two keys:

Airports:	The airports that passengers are travelling between (both directions)

Class:	The class of the ticket.

Examples

library(tsibble)
ansett
#> # A tsibble: 7,407 x 4 [1W]
#> # Key:       Airports, Class [30]
#>        Week Airports Class    Passengers
#>      <week> <chr>    <chr>         <dbl>
#>  1 1989 W28 ADL-PER  Business        193
#>  2 1989 W29 ADL-PER  Business        254
#>  3 1989 W30 ADL-PER  Business        185
#>  4 1989 W31 ADL-PER  Business        254
#>  5 1989 W32 ADL-PER  Business        191
#>  6 1989 W33 ADL-PER  Business        136
#>  7 1989 W34 ADL-PER  Business          0
#>  8 1989 W35 ADL-PER  Business          0
#>  9 1989 W36 ADL-PER  Business          0
#> 10 1989 W37 ADL-PER  Business          0
#> # … with 7,397 more rows