Australian livestock slaughter

Meat production in Australia for human consumption from Q3 1965 to Q4 2018.

Format

Time series of class tsibble

Source

Australian Bureau of Statistics, catalogue number 7218.0.55.001 tables 1 to 6.

Details

aus_livestock is an quarterly tsibble with one value:

Count:	Number of animals slaughtered.

Each series is uniquely identified by one key:

Animal:	The animal slaughtered.

Examples

library(tsibble)
aus_livestock

#> # A tsibble: 29,364 x 4 [1M]
#> # Key:       Animal, State [54]
#>       Month Animal                     State                        Count
#>       <mth> <fct>                      <fct>                        <dbl>
#>  1 1976 Jul Bulls, bullocks and steers Australian Capital Territory  2300
#>  2 1976 Aug Bulls, bullocks and steers Australian Capital Territory  2100
#>  3 1976 Sep Bulls, bullocks and steers Australian Capital Territory  2100
#>  4 1976 Oct Bulls, bullocks and steers Australian Capital Territory  1900
#>  5 1976 Nov Bulls, bullocks and steers Australian Capital Territory  2100
#>  6 1976 Dec Bulls, bullocks and steers Australian Capital Territory  1800
#>  7 1977 Jan Bulls, bullocks and steers Australian Capital Territory  1800
#>  8 1977 Feb Bulls, bullocks and steers Australian Capital Territory  1900
#>  9 1977 Mar Bulls, bullocks and steers Australian Capital Territory  2700
#> 10 1977 Apr Bulls, bullocks and steers Australian Capital Territory  2300
#> # … with 29,354 more rows