NYC Citi Bike trips

A sample from NYC Citi Bike usage of 10 bikes throughout 2018. The data includes event data on each trip, including the trip's start and end times and locations. The customer's gender, birth year and bike usage type is also available.

Format

Time series of class tsibble

Source

Citi Bike NYC, https://www.citibikenyc.com/system-data

Details

nyc_bikes is a tsibble containing event data, the events include these details:

start_time:	The time and date when the trip was started.

stop_time:	The time and date when the trip was ended.

start_station:	A unique identifier for the starting bike station.

start_lat:	The latitude of the starting bike station.

start_long:	The longitude of the starting bike station.

end_station:	A unique identifier for the destination bike station.

end_lat:	The latitutde of the destination bike station.

end_long:	The longitude of the destination bike station.

type:	The type of trip. A "Customer" has purchased either a 24-hour or 3-day pass, and a "Subscriber" has purchased an annual subscription.

birth_year	The bike rider's year of birth.

gender:	The gender of the bike rider.

Each series is uniquely identified by one key:

bike_id:	A unique identifier for the bike.

Examples

library(tsibble)
nyc_bikes

#> # A tsibble: 4,268 x 12 [!] <America/New_York>
#> # Key:       bike_id [10]
#>    bike_id start_time          stop_time           start_station start_lat
#>    <fct>   <dttm>              <dttm>              <fct>             <dbl>
#>  1 26301   2018-02-26 19:11:03 2018-02-26 19:15:40 3186               40.7
#>  2 26301   2018-02-27 07:52:49 2018-02-27 07:58:13 3203               40.7
#>  3 26301   2018-02-27 12:03:27 2018-02-27 12:04:54 3202               40.7
#>  4 26301   2018-02-27 13:53:51 2018-02-27 14:21:04 3638               40.7
#>  5 26301   2018-02-27 14:30:42 2018-02-27 14:33:11 3638               40.7
#>  6 26301   2018-02-27 16:31:04 2018-02-27 16:33:27 3187               40.7
#>  7 26301   2018-02-27 17:37:12 2018-02-27 17:40:49 3638               40.7
#>  8 26301   2018-02-27 17:49:10 2018-02-27 17:55:03 3639               40.7
#>  9 26301   2018-02-27 18:21:57 2018-02-27 18:25:18 3202               40.7
#> 10 26301   2018-02-27 22:08:55 2018-02-27 22:28:42 3638               40.7
#> # … with 4,258 more rows, and 7 more variables: start_long <dbl>,
#> #   end_station <fct>, end_lat <dbl>, end_long <dbl>, type <fct>,
#> #   birth_year <dbl>, gender <fct>