Pelt trading records

Hudson Bay Company trading records for Snowshoe Hare and Canadian Lynx furs from 1845 to 1935. This data contains trade records for all areas of the company.

Format

Time series of class tsibble

Source

Hudson Bay Company

Details

pelt is an annual tsibble with two values:

Hare:	The number of Snowshoe Hare pelts traded.

Lynx:	The number of Canadian Lynx pelts traded.

Examples

library(tsibble)
pelt
#> # A tsibble: 91 x 3 [1Y]
#>     Year  Hare  Lynx
#>    <dbl> <dbl> <dbl>
#>  1  1845 19580 30090
#>  2  1846 19600 45150
#>  3  1847 19610 49150
#>  4  1848 11990 39520
#>  5  1849 28040 21230
#>  6  1850 58000  8420
#>  7  1851 74600  5560
#>  8  1852 75090  5080
#>  9  1853 88480 10170
#> 10  1854 61280 19600
#> # ℹ 81 more rows