Half-hourly electricity demand for Victoria, Australia

vic_elec is a half-hourly tsibble with three values:

Demand:	Total electricity demand in MW.

Temperature:	Temperature of Melbourne (BOM site 086071).

Holiday:	Indicator for if that day is a public holiday.

Format

Time series of class tsibble.

Source

Australian Energy Market Operator.

Details

This data is for operational demand, which is the demand met by local scheduled generating units, semi-scheduled generating units, and non-scheduled intermittent generating units of aggregate capacity larger than 30 MW, and by generation imports to the region. The operational demand excludes the demand met by non-scheduled non-intermittent generating units, non-scheduled intermittent generating units of aggregate capacity smaller than 30 MW, exempt generation (e.g. rooftop solar, gas tri-generation, very small wind farms, etc), and demand of local scheduled loads. It also excludes some very large industrial users (such as mines or smelters).

Examples

library(tsibble)
vic_elec

#> # A tsibble: 52,608 x 5 [30m] <Australia/Melbourne>
#>    Time                Demand Temperature Date       Holiday
#>    <dttm>               <dbl>       <dbl> <date>     <lgl>  
#>  1 2012-01-01 00:00:00  4383.        21.4 2012-01-01 TRUE   
#>  2 2012-01-01 00:30:00  4263.        21.0 2012-01-01 TRUE   
#>  3 2012-01-01 01:00:00  4049.        20.7 2012-01-01 TRUE   
#>  4 2012-01-01 01:30:00  3878.        20.6 2012-01-01 TRUE   
#>  5 2012-01-01 02:00:00  4036.        20.4 2012-01-01 TRUE   
#>  6 2012-01-01 02:30:00  3866.        20.2 2012-01-01 TRUE   
#>  7 2012-01-01 03:00:00  3694.        20.1 2012-01-01 TRUE   
#>  8 2012-01-01 03:30:00  3562.        19.6 2012-01-01 TRUE   
#>  9 2012-01-01 04:00:00  3433.        19.1 2012-01-01 TRUE   
#> 10 2012-01-01 04:30:00  3359.        19.0 2012-01-01 TRUE   
#> # … with 52,598 more rows