Monthly Medicare Australia prescription data

PBS is a monthly tsibble with two values:

Scripts:	Total number of scripts

Cost:	Cost of the scripts in $AUD

Format

Time series of class tsibble

Source

Medicare Australia

Details

The data is disaggregated using four keys:

Concession:	Concessional scripts are given to pensioners, unemployed, dependents, and other card holders

Type:	Co-payments are made until an individual's script expenditure hits a threshold ($290.00 for concession, $1141.80 otherwise). Safety net subsidies are provided to individuals exceeding this amount.

ATC1:	Anatomical Therapeutic Chemical index (level 1)

ATC2:	Anatomical Therapeutic Chemical index (level 2)

Examples

library(tsibble)
#> Registered S3 method overwritten by 'tsibble':
#>   method               from 
#>   as_tibble.grouped_df dplyr
#> 
#> Attaching package: ‘tsibble’
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, union
PBS
#> # A tsibble: 67,596 x 9 [1M]
#> # Key:       Concession, Type, ATC1, ATC2 [336]
#>       Month Concession   Type      ATC1  ATC1_desc ATC2  ATC2_desc Scripts  Cost
#>       <mth> <chr>        <chr>     <chr> <chr>     <chr> <chr>       <dbl> <dbl>
#>  1 1991 Jul Concessional Co-payme… A     Alimenta… A01   STOMATOL…   18228 67877
#>  2 1991 Aug Concessional Co-payme… A     Alimenta… A01   STOMATOL…   15327 57011
#>  3 1991 Sep Concessional Co-payme… A     Alimenta… A01   STOMATOL…   14775 55020
#>  4 1991 Oct Concessional Co-payme… A     Alimenta… A01   STOMATOL…   15380 57222
#>  5 1991 Nov Concessional Co-payme… A     Alimenta… A01   STOMATOL…   14371 52120
#>  6 1991 Dec Concessional Co-payme… A     Alimenta… A01   STOMATOL…   15028 54299
#>  7 1992 Jan Concessional Co-payme… A     Alimenta… A01   STOMATOL…   11040 39753
#>  8 1992 Feb Concessional Co-payme… A     Alimenta… A01   STOMATOL…   15165 54405
#>  9 1992 Mar Concessional Co-payme… A     Alimenta… A01   STOMATOL…   16898 61108
#> 10 1992 Apr Concessional Co-payme… A     Alimenta… A01   STOMATOL…   18141 65356
#> # ℹ 67,586 more rows